src/test/java/com/leetcode/dp/hard/BurstBalloons_312Test.java
package com.leetcode.dp.hard;

import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtensionContext;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.ArgumentsProvider;
import org.junit.jupiter.params.provider.ArgumentsSource;

import java.util.stream.Stream;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.params.provider.Arguments.arguments;

class BurstBalloons_312Test {

  private final BurstBalloons_312 solution = new BurstBalloons_312();

  private static class TestArgumentsProvider implements ArgumentsProvider {
    @Override
    public Stream<? extends Arguments> provideArguments(ExtensionContext context) {
      return Stream.of(arguments(new int[] {3, 1, 5, 8}, 167));
    }
  }

  @ParameterizedTest
  @ArgumentsSource(TestArgumentsProvider.class)
  void maxCoins(int[] nums, int expectedResult) {
    int actualResult = solution.maxCoins(nums);

    assertThat(actualResult).isEqualTo(expectedResult);
  }
}
